refactor: Move ProcessCreationConfig and CoreHookNativeConfig into their own class files
* Since ProcessCreationConfig doesn't belong in the RemoteInjectorConfig class
