+++ src/CoreHook.BinaryInjection/RemoteInjection/RemoteInjector.cs
@@ -12,6 +12,7 @@ using CoreHook.CoreLoad.Data;
 using CoreHook.IPC.Platform;
 using CoreHook.Memory;
 using CoreHook.Memory.Processes;
+using CoreHook.BinaryInjection.ProcessUtils;
 using static CoreHook.BinaryInjection.ProcessUtils.ProcessHelper;
 
 namespace CoreHook.BinaryInjection.RemoteInjection
